Increased test coverage + extracted Event to a separate file
