+++ src/model/__init__.py
@@ -1,3 +1,13 @@
-from .calendar import Calendar, Category, Day, Event, Occurrence
+from .calendar import Calendar
+from .event import Category, Day, Event, Occurrence
+from .utils import this_week, next_week
 
-__all__ = ["Calendar", "Category", "Day", "Event", "Occurrence"]
+__all__ = [
+    "Calendar",
+    "Category",
+    "Day",
+    "Event",
+    "Occurrence",
+    "this_week",
+    "next_week",
+]
