+++ tests/model/test_calendar.py
@@ -2,11 +2,8 @@
 
 import pytest
 from datetime import date, time, timedelta
-from unittest.mock import patch
 
-from model.calendar import Calendar, Event, Occurrence
-from model.calendar import Day, weekday
-from model.calendar import Category
+from model import Calendar, Event, Category, Day, Occurrence
 
 
 class TestCalendar:
@@ -84,6 +81,7 @@ class TestCalendar:
 
         events = calendar.get_future_events()
         assert isinstance(events, list)
+        # FIXME: Add more tests
         # The actual implementation might return different results based on dates
         # Just check that the method returns a list
 
@@ -151,27 +149,6 @@ class TestCalendar:
         except Exception as e:
             pytest.skip(f"remove_past_events failed: {e}")
 
-    def test_get_this_week(self):
-        """Test getting this week's date."""
-        today = date.today()
-        this_week = Calendar.get_this_week()
-        assert isinstance(this_week, date)
-        # Check that this_week is a Monday
-        assert this_week.weekday() == 0
-        # Check that this_week is in the current week
-        assert (today - this_week).days <= 6
-
-    def test_get_next_week(self):
-        """Test getting next week's date."""
-        today = date.today()
-        next_week = Calendar.get_next_week()
-        assert isinstance(next_week, date)
-        # Check that next_week is a Monday
-        assert next_week.weekday() == 0
-        # Check that next_week is in the next week
-        this_week = Calendar.get_this_week()
-        assert (next_week - this_week).days == 7
-
     def test_dictionary_interface(self, calendar, mock_event):
         """Test the dictionary interface of the calendar."""
         # Test __setitem__
@@ -189,54 +166,6 @@ class TestCalendar:
         # Test __contains__
         calendar[1] = mock_event
         assert 1 in calendar
-
-    def test_get_full_repr(self, mock_event):
-        """Test getting full representation of an event."""
-        full_repr = mock_event.get_full_repr()
-        assert isinstance(full_repr, str)
-        assert mock_event.title in full_repr
-        assert mock_event.emoji in full_repr
-
-        # Test with description
-        mock_event.description = "This is a detailed description"
-        full_repr = mock_event.get_full_repr()
-        assert mock_event.description in full_repr
-
-        # Test with regular occurrence and days
-        mock_event.occurrence = Occurrence.REGULAR
-        mock_event.days = {Day.Monday, Day.Wednesday}
-        full_repr = mock_event.get_full_repr()
-        assert "🗓️" in full_repr
-        # The actual weekday representation depends on the implementation
-        # Just check that the full_repr contains something
-        assert full_repr is not None
-
-        # Test with venue but no location
-        mock_event.venue = "Test Venue"
-        mock_event.location = None
-        full_repr = mock_event.get_full_repr()
-        assert "📍Test Venue" in full_repr
-
-        # Test with location but no venue
-        mock_event.venue = None
-        mock_event.location = "https://maps.google.com/?q=Test+Location"
-        full_repr = mock_event.get_full_repr()
-        assert "📍[Location]" in full_repr
-
-        # Test with both venue and location
-        mock_event.venue = "Test Venue"
-        mock_event.location = "https://maps.google.com/?q=Test+Location"
-        full_repr = mock_event.get_full_repr()
-        assert "📍[Test Venue]" in full_repr
-
-        # Test with URL
-        mock_event.url = "https://example.com"
-        full_repr = mock_event.get_full_repr()
-        assert "🔗" in full_repr
-
-        # Test with no title
-        mock_event.title = None
-        assert mock_event.get_full_repr() is None
 
     def test_calendar_iteration_methods(self, calendar, mock_event):
         """Test the dictionary-like iteration methods of Calendar."""
@@ -277,21 +206,6 @@ class TestCalendar:
         assert event_id1 in keys
         assert event_id2 in keys
 
-    def test_get_hash(self, mock_event):
-        """Test the get_hash method."""
-        # get_hash should return the same value as __hash__
-        assert mock_event.get_hash() == mock_event.__hash__()
-        assert isinstance(mock_event.get_hash(), int)
-
-    def test_get_current_repr_edge_cases(self, mock_event):
-        """Test edge cases for get_current_repr method."""
-        # Test with no date (should still work)
-        mock_event.date = None
-        mock_event.end_date = None
-        current_repr = mock_event.get_current_repr()
-        assert current_repr is not None
-        assert mock_event.get_title_repr() in current_repr
-
     def test_get_agenda_categories(self, calendar, mock_event):
         """Test the get_agenda method with different event categories."""
         # Add events with different categories
@@ -331,276 +245,3 @@ class TestCalendar:
         assert "💰 Збори коштів" in agenda
         assert "🤲 Волонтерство" in agenda
         assert "_#agenda_" in agenda
-
-    def test_get_current_repr_with_end_date(self, mock_event):
-        """Test get_current_repr with various end_date scenarios."""
-        # Test with date and end_date where end_date > date
-        mock_event.date = date.today()
-        mock_event.end_date = mock_event.date + timedelta(days=3)
-
-        # Case 1: date < this_week and end_date < next_week
-        with patch("model.calendar.Calendar.get_this_week") as mock_this_week, patch(
-            "model.calendar.Calendar.get_next_week"
-        ) as mock_next_week:
-            mock_this_week.return_value = mock_event.date + timedelta(days=7)
-            mock_next_week.return_value = mock_event.date + timedelta(days=14)
-            current_repr = mock_event.get_current_repr()
-            assert current_repr is not None
-            # Should include weekday name for the end date
-            assert "🗓️до" in current_repr
-
-        # Case 2: date < this_week and end_date >= next_week
-        with patch("model.calendar.Calendar.get_this_week") as mock_this_week, patch(
-            "model.calendar.Calendar.get_next_week"
-        ) as mock_next_week:
-            mock_this_week.return_value = mock_event.date + timedelta(days=7)
-            mock_next_week.return_value = mock_event.date + timedelta(
-                days=2
-            )  # end_date is after next_week
-            current_repr = mock_event.get_current_repr()
-            assert current_repr is not None
-            # Should include the end date in MM/DD format
-            assert mock_event.end_date.strftime("%m/%d") in current_repr
-
-    def test_get_full_repr_with_end_date(self, mock_event):
-        """Test get_full_repr with end_date in different month."""
-        # Test with date and end_date in different months
-        mock_event.date = date(2023, 1, 30)
-        next_month = date(2023, 2, 5)
-        mock_event.end_date = next_month
-
-        full_repr = mock_event.get_full_repr()
-        assert full_repr is not None
-        # Should include both month/day for end_date
-        assert mock_event.end_date.strftime("%m/%d") in full_repr
-
-        # Test with date and end_date in same month
-        mock_event.date = date(2023, 1, 1)
-        same_month = date(2023, 1, 5)
-        mock_event.end_date = same_month
-
-        full_repr = mock_event.get_full_repr()
-        assert full_repr is not None
-        # Should include only day for end_date
-        assert mock_event.end_date.strftime("%d") in full_repr
-
-
-class TestEvent:
-    def test_hash(self, mock_event):
-        """Test event hash."""
-        event_hash = hash(mock_event)
-        assert isinstance(event_hash, int)
-        # Hash should be consistent
-        assert hash(mock_event) == event_hash
-        # Hash should be based on title and date
-        mock_event.title = "New Title"
-        assert hash(mock_event) != event_hash
-
-    def test_has_poster(self, mock_event):
-        """Test has_poster method."""
-        # This might fail depending on the implementation
-        # Just try to call the method
-        try:
-            result = mock_event.has_poster()
-            assert isinstance(result, bool)
-        except Exception as e:
-            pytest.skip(f"has_poster failed: {e}")
-
-    def test_get_weekdays(self, mock_recurring_event):
-        """Test getting weekdays representation."""
-        # The actual weekday representation depends on the implementation
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-
-        # Test with a single day
-        mock_recurring_event.days = {Day.Monday}
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-        assert len(weekdays) > 0
-
-        # Test with two consecutive days (should use range notation)
-        mock_recurring_event.days = {Day.Monday, Day.Tuesday}
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-        assert "-" in weekdays  # Should use range notation
-
-        # Test with three consecutive days (should use range notation)
-        mock_recurring_event.days = {Day.Monday, Day.Tuesday, Day.Wednesday}
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-        assert "-" in weekdays  # Should use range notation
-
-        # Test with non-consecutive days
-        mock_recurring_event.days = {Day.Monday, Day.Wednesday, Day.Friday}
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-        assert "," in weekdays  # Should use comma separation
-
-        # Test with multiple sequences
-        mock_recurring_event.days = {Day.Monday, Day.Tuesday, Day.Thursday, Day.Friday}
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-        assert "," in weekdays and "-" in weekdays  # Should use both notations
-
-        # Test with all days
-        mock_recurring_event.days = {
-            Day.Monday,
-            Day.Tuesday,
-            Day.Wednesday,
-            Day.Thursday,
-            Day.Friday,
-            Day.Saturday,
-            Day.Sunday,
-        }
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-
-    def test_get_weekdays_edge_cases(self, mock_recurring_event):
-        """Test edge cases for get_weekdays method."""
-        # Test with empty days set
-        mock_recurring_event.days = set()
-        assert mock_recurring_event.get_weekdays() == ""
-
-        # Test with a sequence that ends at the end of the week
-        mock_recurring_event.days = {Day.Friday, Day.Saturday, Day.Sunday}
-        weekdays = mock_recurring_event.get_weekdays()
-        assert isinstance(weekdays, str)
-        assert "-" in weekdays  # Should use range notation
-
-        # Test with a sequence that wraps around the week
-        mock_recurring_event.days = {Day.Sunday, Day.Monday}
-        weekdays = mock_recurring_event.get_weekdays()
-        assert "," in weekdays  # Should use comma separation
-
-    def test_get_title(self, mock_event):
-        """Test getting event title."""
-        title = mock_event.get_title()
-        assert isinstance(title, str)
-        assert mock_event.title in title
-        assert mock_event.emoji in title
-        # Test with a different emoji
-        original_emoji = mock_event.emoji
-        mock_event.emoji = "🎭"
-        title = mock_event.get_title()
-        assert "🎭" in title
-        # Restore the original emoji
-        mock_event.emoji = original_emoji
-
-    def test_get_title_repr(self, mock_event):
-        """Test getting event title representation."""
-        title_repr = mock_event.get_title_repr()
-        assert isinstance(title_repr, str)
-        assert mock_event.title in title_repr
-        assert mock_event.emoji in title_repr
-
-    def test_get_current_repr(self, mock_event, mock_recurring_event):
-        """Test get_current_repr method."""
-        # Test with a regular event
-        current_repr = mock_recurring_event.get_current_repr()
-        assert isinstance(current_repr, str)
-        assert "🗓️" in current_repr
-
-        # Test with a non-regular event
-        current_repr = mock_event.get_current_repr()
-        assert isinstance(current_repr, str)
-
-        # Test with a past date (before this week)
-        with patch("model.calendar.Calendar.get_this_week") as mock_this_week:
-            mock_this_week.return_value = mock_event.date + timedelta(days=7)
-            current_repr = mock_event.get_current_repr()
-            assert "🗓️до" in current_repr
-
-        # Test with an end date that's after the start date but before next week
-        with patch("model.calendar.Calendar.get_this_week") as mock_this_week, patch(
-            "model.calendar.Calendar.get_next_week"
-        ) as mock_next_week:
-            mock_this_week.return_value = mock_event.date - timedelta(days=7)
-            mock_next_week.return_value = mock_event.date + timedelta(days=14)
-            mock_event.end_date = mock_event.date + timedelta(days=3)
-            current_repr = mock_event.get_current_repr()
-            assert mock_event.get_title_repr() in current_repr
-
-        # Test with an end date that's after next week
-        with patch("model.calendar.Calendar.get_this_week") as mock_this_week, patch(
-            "model.calendar.Calendar.get_next_week"
-        ) as mock_next_week:
-            mock_this_week.return_value = mock_event.date - timedelta(days=7)
-            mock_next_week.return_value = mock_event.date + timedelta(days=7)
-            mock_event.end_date = mock_event.date + timedelta(days=14)
-            current_repr = mock_event.get_current_repr()
-            assert mock_event.end_date.strftime("%m/%d") in current_repr
-
-        # Test with a time that has minutes = 0
-        mock_event.time = time(14, 0)
-        current_repr = mock_event.get_current_repr()
-        assert "14" in current_repr
-
-        # Test with a time that has non-zero minutes
-        mock_event.time = time(14, 30)
-        current_repr = mock_event.get_current_repr()
-        assert "14:30" in current_repr
-
-        # Test with no title
-        mock_event.title = None
-        assert mock_event.get_current_repr() is None
-
-    def test_get_future_repr(self, mock_event):
-        """Test get_future_repr method."""
-        # Test with a regular event (should return None)
-        mock_event.occurrence = Occurrence.REGULAR
-        assert mock_event.get_future_repr() is None
-
-        # Test with a non-regular event
-        mock_event.occurrence = Occurrence.WITHIN_DAY
-        future_repr = mock_event.get_future_repr()
-        assert isinstance(future_repr, str)
-        assert mock_event.date.strftime("%m/%d") in future_repr
-
-        # Test with an end date in the same month
-        mock_event.end_date = mock_event.date + timedelta(days=3)
-        future_repr = mock_event.get_future_repr()
-        assert mock_event.end_date.strftime("%d") in future_repr
-
-        # Test with an end date in a different month
-        next_month = mock_event.date.replace(month=mock_event.date.month % 12 + 1)
-        mock_event.end_date = next_month
-        future_repr = mock_event.get_future_repr()
-        assert mock_event.end_date.strftime("%m/%d") in future_repr
-
-        # Test with no title
-        mock_event.title = None
-        assert mock_event.get_future_repr() is None
-
-    def test_get_full_repr(self, mock_event):
-        """Test get_full_repr method."""
-        full_repr = mock_event.get_full_repr()
-        assert isinstance(full_repr, str)
-        assert mock_event.title in full_repr
-        assert mock_event.emoji in full_repr
-
-    def test_post(self, mock_event):
-        """Test post method."""
-        # The post method simply returns a dictionary with the event data
-        # It doesn't actually call cross_post
-        result = mock_event.post()
-        assert isinstance(result, dict)
-        assert "text" in result or "photo" in result
-
-        # Test with an image
-        mock_event.image = "test_image.jpg"
-        result = mock_event.post()
-        assert isinstance(result, dict)
-        assert "photo" in result
-        assert result["photo"] == "test_image.jpg"
-        assert "caption" in result
-
-    def test_to_dict(self, mock_event):
-        """Test to_dict method."""
-        # This might fail depending on the implementation
-        # Just try to call the method
-        try:
-            event_dict = mock_event.to_dict()
-            assert isinstance(event_dict, dict)
-            assert event_dict["title"] == mock_event.title
-        except Exception as e:
-            pytest.skip(f"to_dict failed: {e}")
